Effects working with function and use local copy of video
Unfortunately, looks like cross origin video won't work with this lib
because the canvas context won't allow getImageData() on cross origin
sources in the canvas.

This commit allows for passing a filter() function to Video component.
This filter function gets a VideoFrameData as it's argument. It is
a simple class abstracting away the access to the pixels. See the
example app's filter function to see how to implement filter()